Type and indentation of output files errors were corrected

+++ DataReader.h
@@ -1,5 +1,3 @@
-#include <iostream>
-#include <stdio.h>
 using namespace std;
 
 class Data{
@@ -7,8 +5,8 @@ class Data{
     private: 
         char* dir; // directory of the info we need
         char** map; // the map where the agent will act
-        int learning_rate; // learning rate
-        int discount; // discount rate
+        float learning_rate; // learning rate
+        float discount; // discount rate
         int n_size; // width of the map
         int m_size; // height of the map
         int num_iterations; // the number os plays that the agent can do
@@ -21,12 +19,11 @@ class Data{
         Data(char **argv);
         char* getDir();
         char** getMap();
-        int getDiscount();
-        int getLearning_rate();
+        float getDiscount();
+        float getLearning_rate();
         int getWidth();
         int getHeight();
         int getNumIterations();
         void WriteQValues(float ***QValue);
         void WritePolicy(float ***QValue);
-        //TODO: function to write logs and results
 };
